export debug information add vcredist status
status.json in export debug information add vcredist status
