+++ packages/neuron-wallet/src/controllers/export-debug.ts
@@ -10,6 +10,7 @@ import { ckbDataPath } from 'services/ckb-runner'
 import NetworksService from 'services/networks'
 import SyncedBlockNumber from 'models/synced-block-number'
 import AddressService from 'services/addresses'
+import redistCheck from 'utils/redist-check'
 
 export default class ExportDebugController {
   #I18N_PATH = 'export-debug-info'
@@ -51,7 +52,7 @@ export default class ExportDebugController {
     const url = NetworksService.getInstance().getCurrent().remote
     const ckb = new CKB(url)
 
-    const [syncedBlockNumber, ckbVersion, tipBlockNumber, peers] = await Promise.all([
+    const [syncedBlockNumber, ckbVersion, tipBlockNumber, peers, vcredist] = await Promise.all([
       new SyncedBlockNumber()
         .getNextBlock()
         .then(n => n.toString())
@@ -66,7 +67,9 @@ export default class ExportDebugController {
         .catch(() => ''),
       ckb.rpc
         .getPeers()
-        .catch(() => [])
+        .catch(() => []),
+      redistCheck()
+        .catch(() => false),
     ])
     const { platform, arch } = process
     const release = os.release()
@@ -85,6 +88,7 @@ export default class ExportDebugController {
         platform,
         arch,
         release,
+        vcredist
       }
     }
     this.archive.append(JSON.stringify(status), {
